Optimized overall code

+++ src/main/java/edu/library/libraryspringboot/service/RenReqService.java
@@ -10,7 +10,7 @@ public interface RenReqService {
 
     int getReqCount(String uId);
 
-    int getAllReqCount();
+    long getAllReqCount();
 
     void removeIncompleteList(String uId);
